+++ pkg/detest/result_test.go
@@ -13,26 +13,26 @@ type structy struct {
 }
 
 func Test_typeOf(t *testing.T) {
-	assert.Equal(t, "string", describeTypeOfValue("const string"))
+	assert.Equal(t, "string", describeTypeOfActualValue("const string"))
 	s := "var string"
-	assert.Equal(t, "string", describeTypeOfValue(s))
-	assert.Equal(t, "*string", describeTypeOfValue(&s))
+	assert.Equal(t, "string", describeTypeOfActualValue(s))
+	assert.Equal(t, "*string", describeTypeOfActualValue(&s))
 	p := &s
-	assert.Equal(t, "**string", describeTypeOfValue(&p))
-	assert.Equal(t, "stringish", describeTypeOfValue(stringish("foo")))
+	assert.Equal(t, "**string", describeTypeOfActualValue(&p))
+	assert.Equal(t, "stringish", describeTypeOfActualValue(stringish("foo")))
 
-	assert.Equal(t, "int", describeTypeOfValue(42))
+	assert.Equal(t, "int", describeTypeOfActualValue(42))
 	i := 42
-	assert.Equal(t, "int", describeTypeOfValue(i))
+	assert.Equal(t, "int", describeTypeOfActualValue(i))
 
-	assert.Equal(t, "uint8", describeTypeOfValue(uint8(4)))
+	assert.Equal(t, "uint8", describeTypeOfActualValue(uint8(4)))
 
-	assert.Equal(t, "[]string", describeTypeOfValue([]string{}))
-	assert.Equal(t, "[4]string", describeTypeOfValue([4]string{}))
+	assert.Equal(t, "[]string", describeTypeOfActualValue([]string{}))
+	assert.Equal(t, "[4]string", describeTypeOfActualValue([4]string{}))
 
-	assert.Equal(t, "structy", describeTypeOfValue(structy{}))
+	assert.Equal(t, "structy", describeTypeOfActualValue(structy{}))
 
-	assert.Equal(t, "map[string]string", describeTypeOfValue(map[string]string{}))
+	assert.Equal(t, "map[string]string", describeTypeOfActualValue(map[string]string{}))
 
-	assert.Equal(t, "nil", describeTypeOfValue(nil))
+	assert.Equal(t, "nil", describeTypeOfActualValue(nil))
 }